feat:(integração com Cloudinary)
- Foi implementado o upload de imagens usando `multer` e `multer-storage-cloudinary`
- Configura credenciais do Cloudinary no router de produtos
- Cria endpoint POST `/` para adição de produtos vinculados a uma imagem
- Cria endpoint PUT `/:id` com suporte à atualização da imagem (incluindo a exclusão da imagem antiga no Cloudinary)
- Cria endpoint DELETE `/:id` que garante que a imagem vinculada seja removida da nuvem ao deletar o produto

+++ routes/products.js
@@ -1,14 +1,42 @@
-import express from "express"
+import express from "express";
 import Product from "../schemas/Product.js";
 import authMiddleware from "../middleware/authMiddleware.js";
+import multer from "multer";
+import cloudinary from "cloudinary";
+import cloudinaryStorage from "multer-storage-cloudinary";
+import dotenv from "dotenv"
 
-const router = express.Router()
+dotenv.config()
+const router = express.Router();
 
-router.post('/',authMiddleware, async (req, res) => {
+cloudinary.v2.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET
+});
+
+const storage = cloudinaryStorage({
+  cloudinary: cloudinary,
+  folder: "lojinha_produtos",
+  allowedFormats: ["jpg", "png", "jpeg", "webp"],
+});
+
+const upload = multer({ storage: storage });
+
+router.post('/',authMiddleware, upload.single('image'), async (req, res) => {
   try{
-    const newProduct = await Product.create(req.body);
+    const productData = req.body
+
+    if (req.file) {
+      productData.imageUrl = req.file.secure_url;
+      productData.imagePublicId = req.file.filename;
+    } else {
+      return res.status(400).json({ error: "A imagem do produto é obrigatória" });
+    }
+    const newProduct = await Product.create(productData);
     res.status(201).json(newProduct);
   } catch (error) {
+    console.error("Erro no terminal:", error);
     res.status(500).json({ error: "Erro ao criar produto", error});
   }
 
@@ -23,20 +51,53 @@ router.get('/', async (req, res) => {
   }
 })
 
-router.put('/:id',authMiddleware, async (req, res)=> {
+router.put('/:id', authMiddleware, upload.single('image'), async (req, res)=> {
+  try {
+    const productData = req.body;
 
-  try{
-    const updatedProduct = await Product.findByIdAndUpdate(req.params.id, req.body,{new: true})
-    return res.status(200).json(updatedProduct)
+    if (req.file) {
+      const produtoAntigo = await Product.findById(req.params.id);
+
+      if (!produtoAntigo) {
+        return res.status(404).json({ error: "Produto não encontrado" });
+      }
+
+      if (produtoAntigo.imagePublicId) {
+        await cloudinary.v2.uploader.destroy(produtoAntigo.imagePublicId);
+      }
+
+      productData.imageUrl = req.file.secure_url || req.file.url || req.file.path;
+      productData.imagePublicId = req.file.filename;
+    }
+
+    const updatedProduct = await Product.findByIdAndUpdate(
+      req.params.id, 
+      productData, 
+      { returnDocument: 'after' }
+    );
+
+    return res.status(200).json(updatedProduct);
+
   } catch (error) {
-    return res.status(500).json({error: "Erro ao atualizar produto"})
+    return res.status(500).json({ error: "Erro ao atualizar produto" });
   }
-})
+});
 
-router.delete('/delete/:id',authMiddleware, async (req, res)=> {
+router.delete('/:id',authMiddleware, async (req, res)=> {
 
   try{
-    await Product.findByIdAndDelete(req.params.id)
+    const produto = await Product.findById(req.params.id);
+
+    if (!produto) {
+      return res.status(404).json({ error: "Produto não encontrado no banco de dados" });
+    }
+
+    if (produto.imagePublicId) {
+      await cloudinary.v2.uploader.destroy(produto.imagePublicId);
+    }
+
+    await Product.findByIdAndDelete(req.params.id);
+
     return res.status(200).json({message: "Produto deletado"})
 
   } catch (error) {
